requirements.txt
ctransformers==0.2.27
sentence-transformers==4.0.2
pinecone-client
langchain==0.3.24
flask
pypdf
python-dotenv
-e .    